# swagger.yaml
...
info:
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  contact:
    name: Amaury
    url: https://amaurydiaz.github.io
    email: amauryuh@gmail.com
  description: Oauth 2.0 api reference
  title: Authentication
  version: 0.1.0
  termsOfService: ""
paths:
  /auth/oauth/authorize:
    get:
      operationId: Authorize
      parameters:
      - in: query
        name: redirect_uri
        required: true
        type: string
      - in: query
        name: client_id
        required: true
        type: string
      - in: query
        name: state
        required: true
        type: string
      - in: query
        name: scope
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: Ok
        "500":
          description: internal server error
          schema:
            type: string
      security: []
      tags:
      - Oauth
    post:
      consumes:
      - application/x-www-form-urlencoded
      operationId: AuthorizeSubmit
      parameters:
      - enum:
        - code
        - token
        in: formData
        name: response_type
        required: true
        type: string
      - in: formData
        name: redirect_uri
        required: true
        type: string
      - in: formData
        name: client_id
        required: true
        type: string
      - in: formData
        name: state
        required: true
        type: string
      - in: formData
        name: scope
        required: true
        type: string
      - in: formData
        name: username
        required: true
        type: string
      - in: formData
        name: password
        required: true
        type: string
      responses:
        "200":
          description: operation success
        "500":
          description: internal server error
          schema:
            type: string
      security: []
      tags:
      - Oauth
  /auth/oauth/providers/{provider}:
    get:
      operationId: Provider
      parameters:
      - in: path
        name: provider
        required: true
        type: string
      - in: query
        name: code
        required: true
        type: string
      - in: query
        name: state
        required: true
        type: string
      responses:
        "500":
          description: internal server error
          schema:
            type: string
      security: []
      tags:
      - Oauth
...
  /auth/session/recovery:
    get:
      operationId: Recovery
      parameters:
      - in: query
        name: token
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: OK
        "500":
          description: internal server error
          schema:
            type: string
      security: []
      tags:
      - Session
    post:
      consumes:
      - application/x-www-form-urlencoded
      operationId: RecoverySubmit
      parameters:
      - format: password
        in: formData
        name: password
        required: true
        type: string
      - in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: internal server error
          schema:
            type: string
      security: []
      tags:
      - Session
...
schemes:
- https
- http
security:
- bearer: []
securityDefinitions:
  bearer:
    in: header
    name: Authorization
    type: apiKey
  scoped:
    in: query
    name: access_token
    type: apiKey
swagger: "2.0"
...